Make the new metrics work with the regular allocator too
