+++ pkg/gameserverallocations/metrics.go
@@ -40,8 +40,10 @@ var (
 	keyStatus             = mt.MustTagKey("status")
 	keySchedulingStrategy = mt.MustTagKey("scheduling_strategy")
 
-	gameServerAllocationsLatency    = stats.Float64("gameserver_allocations/latency", "The duration of gameserver allocations", "s")
-	gameServerAllocationsRetryTotal = stats.Int64("gameserver_allocations/errors", "The errors of gameserver allocations", "1")
+	gameServerAllocationsLatency        = stats.Float64("gameserver_allocations/latency", "The duration of gameserver allocations", "s")
+	gameServerAllocationsRetryTotal     = stats.Int64("gameserver_allocations/errors", "The errors of gameserver allocations", "1")
+	gameServerAllocationsUpdatesLatency = stats.Float64("gameserver_allocations/update_latency", "The duration of gameserver updates", "s")
+	gameServerAllocationsBatchSize      = stats.Int64("gameserver_allocations/batch", "The gameserver allocations batch size", "1")
 
 	stateViews = []*view.View{
 		{
@@ -56,6 +58,20 @@ var (
 			Measure:     gameServerAllocationsRetryTotal,
 			Description: "The count of gameserver allocation retry until it succeeds",
 			Aggregation: view.Distribution(1, 2, 3, 4, 5),
+			TagKeys:     []tag.Key{keyFleetName, keyClusterName, keyMultiCluster, keyStatus, keySchedulingStrategy},
+		},
+		{
+			Name:        "gameserver_allocations_updates_duration_seconds",
+			Measure:     gameServerAllocationsUpdatesLatency,
+			Description: "The distribution of gameserver allocation update requests latencies.",
+			Aggregation: view.Distribution(0, 0.01, 0.025, 0.05, 0.075, 0.1, 0.25, 0.5, 0.75, 1, 2, 3),
+			TagKeys:     []tag.Key{keyFleetName, keyClusterName, keyMultiCluster, keyStatus, keySchedulingStrategy},
+		},
+		{
+			Name:        "gameserver_allocations_batch_size",
+			Measure:     gameServerAllocationsBatchSize,
+			Description: "The count of gameserver allocations in a batch",
+			Aggregation: view.Distribution(1, 2, 3, 4, 5, 10, 20, 50, 100),
 			TagKeys:     []tag.Key{keyFleetName, keyClusterName, keyMultiCluster, keyStatus, keySchedulingStrategy},
 		},
 	}
@@ -154,4 +170,21 @@ func (r *metrics) record() {
 func (r *metrics) recordAllocationRetrySuccess(ctx context.Context, retryCount int) {
 	mt.RecordWithTags(ctx, []tag.Mutator{tag.Upsert(keyStatus, "Success")},
 		gameServerAllocationsRetryTotal.M(int64(retryCount)))
+}
+
+// record the current allocation batch size.
+func (r *metrics) recordAllocationsBatchSize(ctx context.Context, count int) {
+	stats.Record(ctx, gameServerAllocationsBatchSize.M(int64(count)))
+}
+
+// record the gs successful update latency.
+func (r *metrics) recordAllocationUpdateSuccess(ctx context.Context, duration time.Duration) {
+	mt.RecordWithTags(ctx, []tag.Mutator{tag.Upsert(keyStatus, "Success")},
+		gameServerAllocationsUpdatesLatency.M(duration.Seconds()))
+}
+
+// record the gs failed update latency.
+func (r *metrics) recordAllocationUpdateFailure(ctx context.Context, duration time.Duration) {
+	mt.RecordWithTags(ctx, []tag.Mutator{tag.Upsert(keyStatus, "Failure")},
+		gameServerAllocationsUpdatesLatency.M(duration.Seconds()))
 }
