+++ site/content/en/docs/Guides/metrics.md
@@ -49,8 +49,10 @@ Follow the [Google Cloud Monitoring installation steps](#google-cloud-monitoring
 |-------------------------------------------------------|---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|-----------|
 | agones_gameservers_count                              | The number of gameservers per fleet and status                                                                                                                                              | gauge     |
 | agones_gameserver_allocations_duration_seconds        | The distribution of gameserver allocation requests latencies                                                                                                                                | histogram |
-| agones_gameserver_allocations_retry_total             | The count of gameserver allocation retry until it succeeds | histogram |
-| agones_gameserver_creation_duration                   | The time gameserver takes to be created in seconds  | histogram | 
+| agones_gameserver_allocations_retry_total             | The count of gameserver allocation retry until it succeeds                                                                                                                                  | histogram |
+| agones_gameserver_allocations_updates_duration_seconds| The distribution of gameserver allocation update requests latencies                                                                                                                         | histogram |
+| agones_gameserver_allocations_batch_size              | The count of gameserver allocations in a batch                                                                                                                                              | histogram |
+| agones_gameserver_creation_duration                   | The time gameserver takes to be created in seconds                                                                                                                                          | histogram | 
 | agones_gameservers_total                              | The total of gameservers per fleet and status                                                                                                                                               | counter   |
 | agones_gameserver_player_connected_total              | The total number of players connected to gameservers (Only available when [player tracking]({{< relref "player-tracking.md" >}}) is enabled)                                                | gauge     |
 | agones_gameserver_player_capacity_total               | The available capacity for players on gameservers (Only available when [player tracking]({{< relref "player-tracking.md" >}}) is enabled)                                                   | gauge     |
